Issue 123

+++ app/Http/Controllers/Admin/NotificationsController.php
@@ -15,7 +15,25 @@ class NotificationsController extends Controller
      */
     public function index()
     {
-        return view('admin.notifications.index')->with('notifications', Notification::all());
+        $field = isset($_GET['field']) ? trim($_GET['field']) : "";
+        $dir = isset($_GET['dir']) ? trim($_GET['dir']) : "asc";
+        $notifications = null;
+        
+        if ($field!="") {
+            $notifications = Notification::orderBy($field, $dir)->get();
+        }
+        else {
+            $notifications = Notification::all();
+        }
+
+        $data = [
+            'notifications' => $notifications,
+            'order'  => [
+                'field' => $field,
+                'dir' => $dir
+            ]
+        ];
+        return view('admin.notifications.index')->with('data', $data);
     }
 
     /**
@@ -96,12 +114,17 @@ class NotificationsController extends Controller
             $sids = $data['sids'];
             $obj_ids = explode(",", $sids);
 
+            if (count($obj_ids) ==0 ) {
+                session()->flash('error', 'Nothing has been selected!');
+                return redirect()->route('admin.students.index');
+            }
+
 			foreach ($obj_ids as $id) {
                 $obj = Notification::find($id);
                 $obj->delete();				
 			}
 			            
-            $request->session()->flash('success', 'You have modified information!');
+            $request->session()->flash('success', 'You have removed notifications!');
 			// $url = '';
 			
 			// if (isset($this->request->get['page'])) {
@@ -117,7 +140,7 @@ class NotificationsController extends Controller
             // $this->redirect(HTTPS_SERVER . 'index.php?route=cms/notifications&token=' . $this->session->data['token'] . $url);
             return redirect()->route('admin.notifications.index');
 		}
-        $request->session()->flash('error', 'No id is selected!');
+        $request->session()->flash('error', 'Nothing has been selected!');
         return redirect()->route('admin.notifications.index');
     }
